todo-multiuser-backend/requirements-supabase.txt
fastapi
uvicorn
supabase
python-dotenv
pyjwt
bcrypt
python-multipart
email-validator